+++ tools/nvqpp-v1/nvqpp_diagnostics.h
@@ -8,6 +8,7 @@
  ******************************************************************************/
 #pragma once
 
+#include "nvqpp_args.h"
 #include "clang/Basic/DiagnosticOptions.h"
 #include "clang/Frontend/ChainedDiagnosticConsumer.h"
 #include "clang/Frontend/CompilerInvocation.h"
@@ -17,8 +18,7 @@
 #include "clang/Tooling/Tooling.h"
 #include "llvm/ADT/IntrusiveRefCntPtr.h"
 
-#include "nvqpp_args.h"
-
+// TODO: documentation
 namespace cudaq {
 using DiagnosticsEngine = clang::DiagnosticsEngine;
 
@@ -30,7 +30,6 @@ using LlvmCntPtr = llvm::IntrusiveRefCntPtr<T>;
 //
 struct DiagnosticsOptions : LlvmCntPtr<clang::DiagnosticOptions> {
   using Base = LlvmCntPtr<clang::DiagnosticOptions>;
-
   DiagnosticsOptions(ArgvT argv)
       : Base(clang::CreateAndPopulateDiagOpts(argv)) {}
 };
@@ -40,13 +39,9 @@ struct DiagnosticsOptions : LlvmCntPtr<clang::DiagnosticOptions> {
 //
 struct DiagnosticsBuffer {
   using Base = clang::TextDiagnosticBuffer;
-
   explicit DiagnosticsBuffer() : buffer(new Base()) {}
-
   Base *get() { return buffer.get(); }
-
   void flush(DiagnosticsEngine &engine) { buffer->FlushDiagnostics(engine); }
-
   std::unique_ptr<Base> buffer;
 };
 
@@ -55,12 +50,9 @@ struct DiagnosticsBuffer {
 //
 struct DiagnosticsPrinter {
   using Base = clang::TextDiagnosticPrinter;
-
   explicit DiagnosticsPrinter(DiagnosticsOptions &opts, const std::string &path)
       : printer(new Base(llvm::errs(), opts.get())) {}
-
   Base *get() { return printer.get(); }
-
   std::unique_ptr<Base> printer;
 };
 
@@ -72,21 +64,17 @@ static LlvmCntPtr<clang::DiagnosticIDs> makeIds() {
 }
 
 struct BufferedDiagnostics {
-
   explicit BufferedDiagnostics(llvm::ArrayRef<const char *> argv)
       : opts(argv), buffer(), engine(makeIds(), opts, buffer.get(), false) {
     clang::ProcessWarningOptions(engine, *opts, /* ReportDiags */ false);
   }
-
   void flush() { buffer.flush(engine); }
-
   DiagnosticsOptions opts;
   DiagnosticsBuffer buffer;
   DiagnosticsEngine engine;
 };
 
 struct ErrorsDiagnostics {
-
   explicit ErrorsDiagnostics(llvm::ArrayRef<const char *> argv,
                              const std::string &path)
       : opts(argv), printer(opts, path),
@@ -109,5 +97,4 @@ struct ErrorsDiagnostics {
   DiagnosticsPrinter printer;
   DiagnosticsEngine engine;
 };
-
 } // namespace cudaq
