+++ tools/nvqpp-v1/nvqpp_driver.cpp
@@ -10,7 +10,7 @@
 #include "cudaq/Optimizer/Dialect/CC/CCDialect.h"
 #include "cudaq/Optimizer/Dialect/Quake/QuakeDialect.h"
 #include "cudaq/Optimizer/Support/Verifier.h"
-#include "nvqpp_fe_actions.h"
+#include "frontend_actions.h"
 #include "nvqpp_flag_configs.h"
 #include "nvqpp_options.h"
 #include "clang/CodeGen/BackendUtil.h"
@@ -299,7 +299,16 @@ static std::unique_ptr<clang::FrontendAction> createFrontendAction(
       cudaqArgs.hasArg(cudaq::nvqpp::options::OPT_enable_mlir);
   const std::string cudaqOptPipeline =
       Driver::constructCudaqOptPipeline(cudaqArgs);
-
+  // Check if user just request quake emit action
+  // Note: This is unique to NVQ++ since clang::frontend only support emit bc or
+  // llvm.
+  const bool emitQuakeAction =
+      cudaqArgs.hasArg(cudaq::nvqpp::options::OPT_emit_quake);
+  if (emitQuakeAction) {
+    return std::make_unique<CudaQAction<clang::EmitLLVMAction>>(
+        ci, module, context, cxxMangled, cudaq::nvqpp::OutputGen::EmitMlir,
+        cudaqOptPipeline);
+  }
   switch (act) {
   case (clang::frontend::ActionKind::ASTDump):
     // AST action: no need to invoke CUDAQ
@@ -307,6 +316,11 @@ static std::unique_ptr<clang::FrontendAction> createFrontendAction(
   case (clang::frontend::ActionKind::ASTPrint):
     // AST action: no need to invoke CUDAQ
     return std::make_unique<clang::ASTPrintAction>();
+  case (clang::frontend::ActionKind::EmitBC):
+    return mlirMode ? std::make_unique<CudaQAction<clang::EmitBCAction>>(
+                          ci, module, context, cxxMangled,
+                          cudaq::nvqpp::OutputGen::EmitLlvm, cudaqOptPipeline)
+                    : std::make_unique<clang::EmitBCAction>();
   case (clang::frontend::ActionKind::EmitLLVM):
     return mlirMode ? std::make_unique<CudaQAction<clang::EmitLLVMAction>>(
                           ci, module, context, cxxMangled,
@@ -451,10 +465,22 @@ int Driver::execute() {
   if (comp && !comp->containsError()) {
     llvm::SmallVector<std::pair<int, const clang::driver::Command *>, 4>
         failing;
-    // bool quakeRun = false;
     std::vector<std::string> objFilesToMerge;
     std::vector<std::string> sourceFiles;
     for (auto &Job : comp->getJobs()) {
+      // Collect the list of input source files (i.e., .cpp files)
+      // We use this to detect object files generated by quake/MLIR pipeline.
+      if (Job.getSource().getKind() ==
+          clang::driver::Action::ActionClass::AssembleJobClass) {
+        for (const auto &input : Job.getInputInfos()) {
+          if (input.isFilename()) {
+            llvm::SmallString<128> inputFile =
+                llvm::sys::path::filename(input.getFilename());
+            sourceFiles.emplace_back(inputFile);
+          }
+        }
+      }
+
       const clang::driver::Command *failingCommand = nullptr;
       if (Job.getSource().getKind() ==
           clang::driver::Action::ActionClass::LinkJobClass) {
@@ -521,8 +547,11 @@ int Driver::execute() {
         newLinkArgs.insert(newLinkArgs.end(), makeArgStringRef(rpathDir));
         Job.replaceArguments(newLinkArgs);
       } else {
+        // Not a link job, i.e., frontend (cc1) job
         auto currentArgs = Job.getArguments();
         const bool libMode = [&]() {
+          if (clOptions.hasArg(cudaq::nvqpp::options::OPT_emit_quake))
+            return false;
           // With target
           if (clOptions.hasArg(cudaq::nvqpp::options::OPT_target)) {
             // Use target setting (override CL options)
@@ -557,17 +586,11 @@ int Driver::execute() {
 
         Job.replaceArguments(currentArgs);
       }
-      if (Job.getSource().getKind() ==
-          clang::driver::Action::ActionClass::AssembleJobClass) {
-        for (const auto &input : Job.getInputInfos()) {
-          if (input.isFilename()) {
-            llvm::SmallString<128> inputFile =
-                llvm::sys::path::filename(input.getFilename());
-            sourceFiles.emplace_back(inputFile);
-          }
-        }
-      }
 
+      // Handle target backend compilation:
+      // we piggyback the main compile job and swap out the input and output
+      // args to do backendConfig.cpp compilation. This will ensure that it is
+      // compiled with the same settings (e.g., -fPIC)
       if (Job.getSource().getKind() ==
               clang::driver::Action::ActionClass::AssembleJobClass &&
           !targetConfig.empty() && targetPlatformExtraArgs.genTargetBackend) {
